Fix pom.xml
The template for the pom.xml for this repository was copied from
internal repositories which have a different set of dependencies and
other configuration. Because this repository is a member of the
rspace-os organisation, it should have a pom.xml which is similar to the
other repositories in the organisation so that it doesn't depend on any
internal code.